Updates to messaging behaviour
This commit includes a collection of interrelated updates that adjust how
the conversation between the wizard and user is handled. The original
version of the app allows the wizard to send messages at any time,  to
send multiple consecutive messages, and to advance the current step back or
forward at any point.

The previous version in this branch modified this to force the wizard and
user to exchange a minimum number of messages before the "Next" button
would become enabled, while the "Previous" button was still always available.

This commit implements an extended version of this constraint in addition to
various other changes:

- alternating turns: the wizard can only send a message in response to a
  message from the user (except at the start of a conversation)
- only one utterance/message per turn is allowed
- the input box and "Send" button are disabled at the appropriate times
  to enforce this
- a new ACTION message type is added to allow the chat app to be jumped to
  a specific step
- the "Next" step button is now enabled based on the number of message
  exchanges (an utterance from each user), with the minimum value defined
  in the "min_message_exchange" array in the JSON data loaded by the app
- the "Next" button is additionally only enabled when it is the wizard's
  turn to send a message
- some linting/indentation fixes