Extend Message objects with an interactionTime field
The backend returns InteractionResponse protos to the WoZ app, and while
these contain a protobuf Timestamp field called "time", it doesn't seem
to be getting populated correctly and all the message timestamps end up
being useless (seconds are zeroed but nanos seem to be populated??).

To work around this for the moment, instead take the "interaction_time"
field (also a Timestamp) from the OutputInteraction contained in the
InteractionResponse, it should be a very similar value.