Rework "wait for LLM response" feature
The previous method of displaying a "Please wait while a response is
generated" message while waiting for the LLM API response just set the
text in the message textarea. However this is a bit hacky and could result
in useless messages being logged to the database.

Instead the message is now shown in a separate text widget, replacing the
previous text indicating if the message could be edited or not. That is
no longer necessary as "system" messages will be sent automatically. The
new show_llm_wait_message property is set when an LLM request is submitted
and cleared when any kind of response is received. While it is set, the
"Please wait..." text is displayed.